requirements.txt
pandas
numpy
pytweening
matplotlib
seaborn
av
pillow
